Edits Solution interface so the run function returns a BigInteger. This is due to the possibility of large numbers for future questions. Also, implements Solution6 completely.

+++ ProjectEuler/Program.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -18,7 +19,7 @@ namespace ProjectEuler
             // Question number from argument
             int questionNum = 0;
             // Answer to question
-            int answer;
+            BigInteger answer;
             // Solution object
             Solution s;
             // Stopwatch object for timing
