+++ ProjectEuler/Solution.cs
@@ -1,12 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 
 namespace ProjectEuler
 {
     interface Solution
     {
-        int run();
+        BigInteger run();
     }
 }
